update calculating reward

+++ programs/staking/src/instructions/create_staking.rs
@@ -11,7 +11,7 @@ use crate::state::StakingInfo;
 pub fn create_staking_pool(
     ctx: Context<CreateStakingPool>,
     max_token_amount_per_address: u64,
-    interest_rate: f64,
+    interest_rate: u16,
     start_time: i64,
     end_time: i64,
 ) -> Result<()> {
